refactor(training): infrastructure/models decomposto por agregado (Fase 5.5)
Substitui `src/training/infrastructure/models.py` (monolítico, 456
linhas, 13 Django ORM models) por pacote `infrastructure/models/` com
8 submódulos por agregado + `__init__.py` que re-exporta toda a surface.

Submódulos (um por agregado):
- sessions.py       → TrainingSessionModel + SessionObjectiveModel
- blocks.py         → SessionBlockModel
- wellness.py       → WellnessPreModel + WellnessPostModel
- attendance.py     → AttendanceRecordModel
- execution.py      → ExecutionRecordModel
- communication.py  → FeedbackThreadModel + AttentionQueueItemModel
                      + RecommendationModel
- eligibility.py    → AthleteIneligibilityDeclarationModel
- planning.py       → MesocycleModel + MicrocycleModel

Zero-gap preservado (ORM e registry do Django):
- Nome do pacote mantido como `models/` (não `orm/`) para preservar o
  caminho `<app>.models` esperado pelas migrations (0001..0006) e pelo
  auto-discovery do Django.
- `app_label = "training"` mantido em todos os 13 modelos.
- `training/models.py` (raiz) continua importando via
  `from training.infrastructure.models import ...` — os re-exports em
  `__init__.py` garantem que Django registre todos os modelos.
- Nenhum db_table/index/constraint alterado (verificado com
  `makemigrations training --dry-run --check` → "No changes detected").
- Suite training: 375 passed, 19 skipped (idêntico a Fase 5.4).

Motivação: o arquivo acumulava 456 linhas com 13 agregados Django
ORM. A decomposição casa com o layout já adotado em `domain/entities/`
(Fase 5.3) e `infrastructure/repository/` (Fase 5.4), uniformizando a
navegação por agregado em todas as camadas.
